Replace MockService for PbftNode unit tests
Replaces the existing `MockService` with a new one that, instead of just
doing nothing or returning hard-coded values, logs calls to its methods
and allows for greater flexibility of return values. Currently, it only
allows for configuration of the return values for the `get_settings` and
`summarize_block` methods, since these are the only methods where this
functionality is needed for the planned tests.

Signed-off-by: Logan Seeley <seeley@bitwise.io>